feat(drawing): Rebuilt component with react-native-sketch-canvas
This is a temporary implementation as react-native-sketch-canvas is abandoned. The current package is actually a fork of @sourcetoad author that has added some basic compatibility with the latest React Native. But the fork is not really reliable as it is maintained by one person and there is no clear plans/intentions to maintain it further.

https://mindlogger.atlassian.net/browse/M2-4100