[M2-5044] [M2-4793] 'outside of' rule for slider works incorrectly (#635)
This PR fixes the logic within `AnswerValidator.isOutsideOfValues` which was using `ConditionalLogicModel.isBetweenValues` instead of `ConditionalLogicModel.isOutsideOfValues`.

I also took the opportunity to update the logic for the `Next` button because I noticed while testing that the respondent may discover the logic for how to unlock some steps that they may other not have, just by playing with the slider. This brings the mobile app in line with web (see https://github.com/ChildMindInstitute/mindlogger-web-refactor/pull/358).