hotfix: Fix infinite deep linking redirection on Android (M2-6600) (#764)
I've included an explicit `android:pathPrefix` attribute in the deep linking intent filter in the Android manifest file. Right now we only support one path for deep linking: `/password-recovery/...`. By explicitly proving this value at build time, it will prevent the OS from launching the app for any other paths on an associated domain.

I'm not sure if this is a scalable solution, but it should work for our current set of paths