fix: remove manual `fontFamily` overrides
No component should be performing the `IS_IOS` logic to determine
which font to use, that was already done in Tamagui config and
components should rely on that, which also provides language-specific
font. So removed those `fontFamily` overrides in `ToastConfig` and
`Splash`.

In `Splash`, the RN-provided Text component was being used to display
the version text. It may inherit Tamagui styling by default, but to
safe, it made sense to me to use the shared Tamagui-based `Text`
component. Also added missing translation for "Version" word.