fix: ensure fonts are properly registered
There are many pitfalls in adding fonts to a React Native app, both on
iOS and Android. This commit addresses those.

**For iOS:**
We have numerous targets in the Xcode project, one for each
environment. Each of those has their own `.plist` file and configuration
settings in `project.pbxproj`, and needed to have the new fonts
accounted for in each place.

**For Android:**
Android is more particular about specific font styles, and per
Tamagui docs, you have to explictly define which font face should be
used for each specific weight. In making this change to
`tamagui.config.js`, it made sense to include all possible font weights
in the project to align as closely as possible with the font weights
actually in use.