feat: Greek font change (M2-8638) (#946)
* feat: Add font files

* feat: RN asset updates

* feat: Add useFontLanguage hook

* feat: Implement FontLanguageProvider

* fix: Correct state variable value assignment

* refactor: simplify `useFontLanguage` hook
I don't see the need for `useState` + `useEffect`, can just return a
variable based directly to the current language, which is reactive.

* chore: remove reference to unused font token

* fix: remove manual `fontFamily` overrides

No component should be performing the `IS_IOS` logic to determine
which font to use, that was already done in Tamagui config and
components should rely on that, which also provides language-specific
font. So removed those `fontFamily` overrides in `ToastConfig` and
`Splash`.

In `Splash`, the RN-provided Text component was being used to display
the version text. It may inherit Tamagui styling by default, but to
safe, it made sense to me to use the shared Tamagui-based `Text`
component. Also added missing translation for "Version" word.

* fix: ensure fonts are properly registered

There are many pitfalls in adding fonts to a React Native app, both on
iOS and Android. This commit addresses those.

**For iOS:**
We have numerous targets in the Xcode project, one for each
environment. Each of those has their own `.plist` file and configuration
settings in `project.pbxproj`, and needed to have the new fonts
accounted for in each place.

**For Android:**
Android is more particular about specific font styles, and per
Tamagui docs, you have to explictly define which font face should be
used for each specific weight. In making this change to
`tamagui.config.js`, it made sense to include all possible font weights
in the project to align as closely as possible with the font weights
actually in use.

* fix: correct default font styling on shared `Text`

Set up default styling on shared `Text` component that can be overridden
by passed props and/or `style` prop. (Using `styled` turns out to be
too prescriptive for instances of attribute-based styling.)

* fix: use new title component based on Tamagui

The navigation header was using a text component based on RN Text, not
our shared `Text` component based on Tamagui. Thus it wasn't picking up
on the font style changes implemented to respond based on selected
language. Fixed this by using a component for the `headerTitle` option
rather than using `headerTitleStyle`.

* fix: apply font to `MarkdownView`

* fix: always use shared `Text` component

---------

Co-authored-by: farmerpaul <paul.hh@metalab.com>