fix: Mobile app fixes (M2-8800, M2-8801) (#952)
* fix: Prohibit font shorthand props on `Text`

Apparently Tamagui provides shorthand props for font styling that were
not being accounted for in our shared `Text` component. For consistency
and predictable handling of default values, updated `Text` to exclude
those from the valid props list.

* fix: disable redundant back button

On Android, when a default `headerTitle` component is provided, it
changes the default value of `headerBackVisible` to `true`. The back
button styling is always provided manually and should never be shown. So
made it explicitly `false`.