feat: Flesh out all remaining sub-step logic
Also fixed some of the conditional logic with animation direction
calculation. Move the animation direction logic out of an effect because
it doesn't actually change state (changes a Reanimated shared value),
and needs to be updated immediately during render to get the correct
value of the current and previous sub-steps.