feat: Allow item types to provide own scrollers
We want to manage scrolling using a `FlatList` on the 3rd step of the
EHR item type, rather than use the default scrolling mechanism currently
provided. So added a way to disable that for this item type, reusing it
on most sub-steps but then using our `FlatList` on the 3rd step instead.