feat: Implement health system search UI
Added paginated (infinite scroll) support of retrieving health system
search results, implementing as much of the querying logic in the hook
as possible, rather than complicating the UI component.

Created a hook `useOneUpHealthBaseQuery` to properly retrieve the error
message returned by 1UpHealth, rather than `useBaseQuery` which only
recognizes the error message structure of our own BE. We need to use
this error message to determine when specifically the access token needs
to be refreshed. Added token refresh support for up to 3 attempts. Did
not implement other nuanced error state handling, leaving that for an
upcoming ticket.