feat: Add full 1UpHealth respondent flow for consenting to share EHRs (M2-8887) (#974)
* feat: Add `active-assessment` deep link

Extend navigation service to support in-progress activity navigation
state used for active assessment deep link. In doing so, replace
`useInitialRouteNavigation` with `useInitialNavigationState`, which
does a better job of preserving navigation state than the previous
approach.

* feat: Scaffold out `subStep` support

Add basic support for initializing the new `subStep` property for EHR
items, and scaffold out the screens. Sub-step navigation and animation
to follow.

* feat: Add new `usePreviousValue` hook, distinguish

Copied `usePrevious` hook from the web app to this app, but renamed it
to `usePreviousValue` to distinguish it from this app's existing
`usePrevious` hook, which behaves differently. Renamed existing
`usePrevious` hook to `usePreviousRenderedValue` to
distinguish its functionality.

Will use `usePreviousValue` in sub-step animation.

* feat: Add sub-steps navigation

* refactor: Get deep-link logic working correctly

After fixing the Xcode configuration, discovered that the
`active-assessment` deep link wasn't opening correctly. Refactored
slightly and fixed the issue. Created useActiveAssessmentLink hook for
handling the logic for updating the EHR item state when resuming an
active assessment via deep link.

* feat: Flesh out all remaining sub-step logic

Also fixed some of the conditional logic with animation direction
calculation. Move the animation direction logic out of an effect because
it doesn't actually change state (changes a Reanimated shared value),
and needs to be updated immediately during render to get the correct
value of the current and previous sub-steps.

* test: Fix existing tests

* feat: Generate `submit_id` when starting surveys

This change renames the `executionGroupKey` property, which used to only
be present on activity flow assessments, to `submitId` (since that's its
real purpose), and also attaches this property to regular activity
assessments. Generating it when starting (or restarting) any assessment
ensures it's available at any time during the assessment, and thus can
be passed as a parameter to the 1UpHealth token endpoint.

* feat: Add 1UpHealth API infrastructure

* feat: Add support for 1UpHealth `/clinical/${id}`

This endpoint requires the 1UpHealth client ID, so documented this as a
requirement in the README.

* feat: Allow item types to provide own scrollers

We want to manage scrolling using a `FlatList` on the 3rd step of the
EHR item type, rather than use the default scrolling mechanism currently
provided. So added a way to disable that for this item type, reusing it
on most sub-steps but then using our `FlatList` on the 3rd step instead.

* feat: Implement health system search UI

Added paginated (infinite scroll) support of retrieving health system
search results, implementing as much of the querying logic in the hook
as possible, rather than complicating the UI component.

Created a hook `useOneUpHealthBaseQuery` to properly retrieve the error
message returned by 1UpHealth, rather than `useBaseQuery` which only
recognizes the error message structure of our own BE. We need to use
this error message to determine when specifically the access token needs
to be refreshed. Added token refresh support for up to 3 attempts. Did
not implement other nuanced error state handling, leaving that for an
upcoming ticket.

* fix: Fix deep link recognition

1UpHealth appends query parameters to the deep link URL, so needed to
adjust how we recognize the `active-assessment` deep link. Also
removed the superfluous `/` in the reconstruction of the deep link for
logging, etc.

* feat: Add `allowedEhrIngest` to answer payload

* fix: Minor UI improvements/fixes

Remove warnings about box shadow performance by adding a background
colour to image thumbnails. Address other warning about incorrect use
of Tamagui tokens in style definition. Minor visual improvements.

* feat: Add no results empty state

* fix: Minor UI/UX improvements

* ci: Fix TSC errors

* fix: Android UI/UX fixes

* tests: Fix existing unit tests

* test: Add tests for `useSubSteps`

* test: Add tests for `useActiveAssessmentLink`

* test: Make test make more sense