fix: Align safe area container bg w/ headers
On iOS, safe area now needs to have its background color aligned with
the `headerStyle.backgroundColor` screen option. Unfortunately,
React Navigation fails to provide developers with a "getter" for this
option, so we have to track it separately in Redux.

There are only two screens that override the background colour from
the default: the applets list view, and the single, tabbed applet view.