fix: Fix header backgrounds & top margin on iOS (#983) (M2-9329)
The introduction of banners resulted in a layout & background colour issue on the assessment screens, an oversight on my part. There was both too big of a top gap, and the topmost header area behind the iPhone hardware no longer matched the background of the rest of the page. The background mismatch was happening on a handful of other screens during the assessment, including the intermediate answers submission screen during a flow, and the summary screen.

This PR fixes it. It's only relevant to iOS devices, especially iPhone.