M2-9405: Simplify SketchCanvas path logic
This feels deceptively simple. I was hesitant to do this refactoring at first because the MAX_POINTS_PER_LINE (set to 50) seemed intentional. However, I did some digging and was unable to find any reasoning behind it.

I tested this on Android versions 13 through 16, as well as iOS 18.3, and it seems to work even for really long paths