M2-9409: Move explicit font size and line height to body property
Because of the way the markdown tree is laid out, headings contain text nodes. Therefore, style rules defined in the `text` property will seem to "override" rules defined in a more specific property (e.g. `heading1`). However, what's happening is just the rules applying as intended.

I have moved the explicit fontSize and lineHeight rules to the `body` property, as this is probably what we want instead