fix: Applet list refresh/loading state (M2-9430)
Something related to navigation transitions since Banners were
introduced caused the Applets list screen to not render for a period
the first time you navigate to it after logging in. Fixed using a bit of
a kluge to prompt another render immediately upon navigation. Also
added a proper loading state spinner, only shown the first time cache
is hydrated (not when doing a pull-to-refresh, which has its own
loading spinner).