feat: Improve Spinner, use more consistently
Added overlay and transition support to the Spinner for use in multiple
places. There are still some places where it could be used, may tidy
those up later.

The Loading component was never in use, so it was removed.

While testing Spinner in the PasswordRecoveryScreen, discovered and
fixed multiple bugs:

- The ForgotPassword screen that is reached from password recovery
  was not accessible if you had an active session. Although this is an
  edge case, it's annoying enough during testing that I decided to fix
  it by moving ForgotPassword into a common part of the RootNavigator.
- If the app is fully closed, and you click a password recovery link,
  the app would crash because the useDefaultBanners hook was making
  assumptions about the navigation state. Made it more resilient.